+++ Opencv-VS/src/img_detection.h
@@ -6,8 +6,14 @@
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
 
-#include "histogram.h"
+#include "img_histogram.h"
+#include "sort_functions.h"
 
 int detectCircle(cv::Mat &image);
+
+cv::Mat connectComponentsWithStats(cv::Mat image);
+
+void detectFeaturePointsAKAZE();
+void detectFeaturePointsFFD();
 
 #endif
