+++ Opencv-VS/src/img_filtering.h
@@ -13,12 +13,9 @@ cv::Mat removeLightWithDivision(cv::Mat image, cv::Mat pattern);
 cv::Mat createLightPattern(cv::Mat image);
 
 cv::Mat binarizeImage(cv::Mat image);
-cv::Mat binarizeImageInverse(cv::Mat image);
-cv::Mat binarizeImageOtsu(cv::Mat image);
-
-cv::Mat connectComponents(cv::Mat image);
-cv::Mat connectComponentsWithStats(cv::Mat image);
 
 cv::Mat sharpen (cv::Mat image);
+
+cv::Mat normalizeColoring(cv::Mat img);
 
 #endif
